gvfs-mount: Handle the ask-question signal
Present questions as a message and a numbered list of choices such
that the user must enter one of the numbers.

For example:
"""
Can't verify the identity of ...
This happens when you log in to a computer the first time.

The identity sent by the remote computer is ... If you want to be
absolutely sure it is safe to continue, contact the system
administrator.
[1] Log In Anyway
[2] Cancel Login
Choice: 1
"""

https://bugzilla.gnome.org/show_bug.cgi?id=728959