http: Use a global SoupSession rather than per-backend
Use a global SoupSession rather than a per-backend so that idle
connections are reused appropriately.  This prevents a process running
out of file descriptors due to connections stuck in a CLOSE_WAIT state.

It also has the side-effect of limiting the maximum number of concurrent
connections because libsoup's max-conns property is respected, fixing
bug 576460.

(Reapplying now that #732783 and #732925 have been fixed.)

https://bugzilla.gnome.org/show_bug.cgi?id=732090