afp: Fix race between writing and reading
The following sequence of events is possible in GVfsAfpConnection:
send_request_unlocked
write_dsi_header_cb
... return to main loop ...
read_dsi_header_cb
write_command_cb

This happens if the server sends its response before the main loop gets
a chance to run since write_command_cb is executed asynchronously as an
idle function.  It causes the job to hang because the request is only
stored in the request hash table in write_command_cb and so the response
is ignored when being processed because it cannot find the corresponding
request.

To fix this, store the request in the hash table before the request is
written.

https://bugzilla.gnome.org/show_bug.cgi?id=733133