afp: Disable Nagle's algorithm
Disable Nagle's algorithm for the connection to the server to improve
performance.

Due to the backend doing a write-write-read sequence, the combination of
Nagle's algorithm and TCP delayed ACKs means that the second write is
delayed until the server's delayed ACK timer times out (40ms on Linux).
This results in each request-response taking far too long causing poor
performance.

This patch reduces the time it takes to duplicate a directory of 10000
small files from 21 minutes to 11 seconds.

https://bugzilla.gnome.org/show_bug.cgi?id=733217