fuse: Use query_info_on_{read,write} if a stream is open
If a stream is open and it supports query_info_on_{read,write}, use that
in preference to querying the file using the path.  This causes the
correct information to be retrieved when g_file_replace is implemented
by writing to a temporary file, rather than the information of the
yet-to-be-replaced file.

https://bugzilla.gnome.org/show_bug.cgi?id=632296