fuse: Close stream on release rather than flush
Before, if an application created, dup()ed the fd and then closed it,
the stream would be closed.  When attempting to write to the original
fd, reopening the stream would fail since it is an error to open an
existing file without either truncating it or appending to it.

Close the stream on release() rather than flush() to fix this since
release() is called only once per open/create rather than for each
close.  Since g_output_stream_flush() is called after every write,
flush() and fsync() are unnecessary.

https://bugzilla.gnome.org/show_bug.cgi?id=632296