mtp: implemented support for LIBMTP_EVENT_STORE_REMOVED
This patch adds support for LIBMTP_EVENT_STORE_REMOVED, so that if a
storage is removed, its entry and those of its children are
automatically removed. For each such entry, delete event is also
emitted.

With this patch, the corresponding storage disappears from Nautilus
as soon as the SD card is unmounted on device. Similarly, if Nautilus
is located in the storage or one of its sub-folders, the window gets
properly closed after the storage is removed.

Signed-off-by: Rok Mandeljc <rok.mandeljc@gmail.com>