MTP: Emit delete events when unmounting.
Although it would seem obvious that a mount disappearing should
invalidate any file references in clients, recent versions of
Nautilus will keep open windows around, showing all the files that
were there at the time of the unmount.

So, clearly we need to emit these events to get it to clean itself
up.