mtp: added support for OBJECT_ADDED event
This patch adds support for OBJECT_ADDED event, which is emitted
when a file or a folder is created on the device. When such an
event is received, its path is determined by first looking up the
parent ID and appending object's name to it; the path and IDs are
added to cache, and create event is emitted, causing clients (for
example Nautilus) to display the new file/folder without having to
refresh the display of parent folder.

Signed-off-by: Rok Mandeljc <rok.mandeljc@gmail.com>