fuse: Pass the correct flags when reopening output stream
If a file is open for reading and writing, ensure to pass the O_APPEND
flag if necessary when reopening the output stream.  This allows a
write-read-write sequence to work properly if the file is opened for
O_APPEND.  The O_TRUNC flag is not passed since this would cause a
write-read-write sequence to truncate the file on the second write.  As
it is, the second write fails with ENOTSUP since this kind of operation
is not supported by GVFS.

https://bugzilla.gnome.org/show_bug.cgi?id=632296