mtp: do_open_for_read() should fail on a directory
This is a fix for bug #729628: the GIO open-read-write-close
fallback for copy/move operation on MTP volumes works correctly
for files, but fails for directories - in particular, it creates
a file with the destination name, then fails with "libmtp error:
Unknown error." This happens because the fallback first attempts
to open the source as a file, and if it succeeds, proceeds to
create destination file and copy the contents via series of
reads/writes.

For directories, do_open_for_read() should fail with
G_IO_ERROR_IS_DIRECTORY, which in tun causes GIO to create the
destination as directory, and then copy its contents over.

Signed-off-by: Rok Mandeljc <rok.mandeljc@gmail.com>