ftp: Also mark connection as unusable on G_IO_IN
There is data waiting on an FTP connection and we didn't request it.
It's not ours and we have no way to clear it. Just makr this connection
as unusable.

Most likely this is the server notifying us about a timeout or other
connection abort, so it's a good idea to treat it as an error anyway.

Also includes a new debug message for when we mark a connection as
unusuable and why.

https://bugzilla.gnome.org/show_bug.cgi?id=591054