metatree: avoid endless looping when the entry is too large
When an application tries to save a larger key-value pair than the size
of the journal, it triggers the journal to be flushed to make space for
the entry and the operation is then retried, but it never fits, and the
loop continues forever.

This patch removes the endless retry loop and retries the operation
only once after the flush. We know that there isn't enough space for
the entry if it fails after the flush.

https://bugzilla.gnome.org/show_bug.cgi?id=637095