MTP: Work around slow large file transfers to Android devices
I've observed that on at least some Android devices, copying large files
is very very slow. I don't know if this hardware or MTP stack specific
but it's a very real phenomenon. Files under 4GB copy at normal speeds,
as do copies from the device to the host at any size.

To avoid this, we can avoid using SendObject and instead take advantage
of the Android Direct I/O extensions.

It may turn out to be desirable to make a similar change for GetObject,
and/or < 4GB transfers.

https://bugzilla.gnome.org/show_bug.cgi?id=736495