Allow cancelling jobs that have sent a reply but not finished
A few job types (mount, unmount and enumerate) can be in a state where
sent_reply is TRUE but finished is FALSE because they override
send_reply.  If the peer connection closes during this period, the
daemon hangs in peer_connection_closed because the job still exists but
cannot be cancelled.  To fix this, allow cancelling jobs that have sent
a reply but not yet finished.

This can be reproduced fairly easily by doing a search on an sftp mount
in Nautilus.

https://bugzilla.gnome.org/show_bug.cgi?id=720860