sftp: Retrieve the username from the secret store
Retrieve the username from the secret store rather than simply using the
default username with the retrieved password.  Without this, the backend
does not correctly recall the login details of a non-default user when
the username is not specified in the URI.

E.g. if my username is "ross" and I stored the login details for "john",
logging in with a URI of sftp://host/path should use john's details
rather than john's password with ross as a username.

If the username is specified directly, e.g. sftp://john@host/path, this
already did work correctly (and should continue to).

https://bugzilla.gnome.org/show_bug.cgi?id=736311