goa: Pass "password" as ID when fetching the password
Passing GoaAccount:id to GetPassword makes no logical sense. It works
because it is ignored by GOA for ownCloud accounts.

The ID is meant as a key when an account might have multiple passwords.
eg., an email account can have separate passwords for IMAP and SMTP,
and the keys can be "imap-password", "smtp-password". This is not the
case for ownCloud and the ID field is ignored by GOA because the key
is always "password". Since ownCloud is the only password-based
account that this volume monitor supports, let's just use "password".

https://bugzilla.gnome.org/show_bug.cgi?id=739078