client: Don't guard against g_object_new returning NULL
The call to g_daemon_mount_new is a call to g_object_new like any
other. There is no point in checking if g_object_new failed to create
an object, because it will always succeed unless we have run out of
virtual address space and in that case the program will be aborted
anyway.

https://bugzilla.gnome.org/show_bug.cgi?id=739075