sftp: Add support for multiple connections
Multiple connections are needed since 7890d2801a7f ("sftp: Implement
push support") and ed826fdf386c ("sftp: Implement pull support"). This
is because up to 2 MiB of data may be in flight, so any "command" sent
(e.g. refreshing the view in Nautilus) gets queued up which makes the
mount unusable on a slow connection when copying at the same time.

This changes the code to support multiple connections but still uses
a single connection.

https://bugzilla.gnome.org/show_bug.cgi?id=747128