ftp: Use TCP_NODELAY
When ftp is layered on top of TLS, it does a write-write-read which
causes a large amount of latency due to the combination of Nagle's
algorithm and delayed ACKs. Use TCP_NODELAY to disable Nagle's algorithm
and prevent this.

The flag is used for both secure and normal connections.  This should
not cause any issues.

https://bugzilla.gnome.org/show_bug.cgi?id=526582