dav: Verify TLS certificates
When mounting a secure webdav share, verify the certificate using the
system certificate store, or if that is not possible, ask the user
whether it is OK.

ssl-strict is enabled for the SoupSession.  If the first connection
attempt fails, the certificate is presented to the user.  If they agree
to the certificate, then it is stored and compared with the certificate
presented on each subsequent connection.  It is an error if the
certificate changes.

https://bugzilla.gnome.org/show_bug.cgi?id=708306