ftp: Prevent segfault when unmounting
If a job is waiting for a connection in
g_vfs_ftp_task_acquire_connection and the backend is force unmounted,
the queue is freed, and the function neither gets a connection nor
sets an error message which causes an issue later.

To fix this, set an error message if the queue is freed.

https://bugzilla.gnome.org/show_bug.cgi?id=653042