sftp: Login to both connections when mounting
Login to the command connection and the data connection when mounting.
Any password entered during the first connection is reused for the
second connection.

Because the data connection is not essential, do not fail if the
connection attempt fails, and introduce a function to determine if a
connection is usable.

https://bugzilla.gnome.org/show_bug.cgi?id=747128