dav: Emit progress callbacks when copying and moving
For progress in applications like Nautilus to be shown correctly, they
need at least one progress callback to be emitted at the end. This is
supposed to be guaranteed according to the GIO documentation. With
server-side copy and move using webdav, emit a single progress callback
at the end with the source file size.

https://bugzilla.gnome.org/show_bug.cgi?id=749354