afc: Remove attempts at thumbnailing photos
iDevices had a list of thumbnails available through afc:/// for photos
stored in the DCIM/ sub-directory, but the code to handle this was
always pretty fragile, relying on the iOS version to know which of a
number of different ways the thumbnails could be called or where they
would be stored.

It's easier to create the thumbnails ourselves from the JPG files, or
even access photos through PTP so that thumbnails are readily
accessible and in a way that doesn't change with each iOS release.

https://bugzilla.gnome.org/show_bug.cgi?id=749639