archive: Handle ARCHIVE_FAILED error
Archive backend is stucked in endless loop currently e.g. if you try to
mount encrypted zip file. It is caused because ARCHIVE_FAILED error is
not handled, when reading data from the archive.

ARCHIVE_FAILED is handled with this patch. Mount job doesn't fail, just
because of unknown file size, but open_for_read job fails if
ARCHIVE_FAILED is returned.

https://bugzilla.gnome.org/show_bug.cgi?id=752366