sftp: Improve error messages on connection setup
Log all the messages from SSH's stderr.

If the remote server does not have sftp setup correctly, it fails with
"subsystem request failed". Report this as "Connection refused".

If the error is unknown, report it as "Connection failed" rather than
"SSH program unexpectedly exited" which shouldn't be exposed to users.

https://bugzilla.gnome.org/show_bug.cgi?id=752403