ftp: Sanitize absolute symlink paths when resolving
The existing code sanitizes relative symlinks but not absolute symlinks
when resolving them. But this can cause issues when looking up the
resolved path (e.g. if it has a trailing slash, the correct information
will not be returned because it will not get the parent directory
correctly).

To fix this, sanitize both absolute and relative paths.

https://bugzilla.gnome.org/show_bug.cgi?id=755303