Use conventional naming for D-Bus session services
This naming is mandatory for the system bus, but is also recommended
for the session bus.

The D-Bus maintainers recommend that all activatable session services'
service files are named according to the bus name, so that any conflict
is resolved in a deterministic way. If the services are in different
directories (precedence levels) the result is the same as it is now:
the higher precedence "wins". If the services are in the same
directory, either one overwrites the other and consistently "wins",
or a packaging system like dpkg prevents co-installation.

If the service files were named differently, it would be possible
to have two implementations for the same name. dbus-daemon resolves
this by choosing one arbitrarily, not necessarily the same one every
time. systemd's kdbus support is more strict (or less concerned with
backwards compatibility), and treats this situation as an error.

Bug: https://bugzilla.gnome.org/show_bug.cgi?id=755760