file monitor: Construct synchronously
g_file_monitor_file() is a sync call anyway, so doing some sync i/o
in the constructor doesn't make a difference, and doing so avoids
races where you would not get change events for operations you do.

We don't actually need to wait for the subscribe event, and proxy
construction in this case does not do any i/o, so we're really only
doing sync i/o to look up the connection, and it should be cached from
the earlier call in g_daemon_file_monitor_file(), so in practice
this should make little difference.

https://bugzilla.gnome.org/show_bug.cgi?id=749317