google: Mark files you can't see on the web as hidden
Files without the parent are currently shown in google backend root.
You can see many files consequently, which are not shown in your google
drive folder on the web. Those files can be shared with you, or you
opened them in some google applications (i.e. google docs). Mark those
files as hidden to see same files as you can see on the web. Also mark
the files as undeletable, because it is not possible to delete them
from the obvious reason.

https://bugzilla.gnome.org/show_bug.cgi?id=756644