proxy volume monitor: Properly handle failure to create a remote proxy
We should finish constructing the innards of the object and not leave
it in an inconsistent state when we hit an error. The other option
would be to litter the rest of the code with NULL checks, but that
would be ugly and prone to errors.

We should also ensure that the reference counting stays consistent with
the non-error paths.

https://bugzilla.gnome.org/show_bug.cgi?id=755805