metadata: Avoid endless recursion when copying meta files
meta_builder_copy is traversing the tree while at the same time possibly
writing to it. This might leads to endless recursion and consequent
segmentation fault. Do the entire copy into a new MetaFile that is not
inserted into the source tree and consequently insert the MetaFile once
the copy is done.

https://bugzilla.gnome.org/show_bug.cgi?id=759341