udisks2: Send unmount notification only once when stopping
Unmount notification is sent for each volume when stopping drive,
which is wrong. It may lead to data loss, because user may unplug
the device after the first notification, before the device is really
stopped. It should be handled similary as an eject operation and
send the notification only once for the drive.

https://bugzilla.gnome.org/show_bug.cgi?id=760523