afc: Don't mount the default AFC service
There's no interesting data for users in afc:/// so it's best to show
iDevices through MTP instead. The only reason we'd want to use this
mount is if we had support for accessing the iTunes database.
Unfortunately, this database needs to be encrypted, and we don't know
how to encrypt it.

Even then, we would probably want the default AFC service to be hidden
so only applications access it, and not users.

https://bugzilla.gnome.org/show_bug.cgi?id=749639