afc: Limit the number of HouseArrest services we start
To avoid being unable to navigate within applications because we ran out
of HouseArrest services, keep track of file operations within each
application, and disconnect from unused ones when trying to access an
unconnected one.

The tracking only covers file operations, and not enumerations, and the
garbage collection algorithm is quite naive, but could easily be
extended.

https://bugzilla.gnome.org/show_bug.cgi?id=676188