gphoto2: Use G_FILESYSTEM_PREVIEW_TYPE_NEVER
PTP/MTP devices can only transfer complete files, which means that
generating a thumbnail can often take multiple minutes, and any
other operation is blocked until this is done. In practice this can
often prevent meaningful use of the device. So change the hint to
applications, advising them it is a bad idea to generate thumbnails
for these devices.

https://bugzilla.gnome.org/show_bug.cgi?id=765843