Always add mount prefix in cached fuse paths
If a fuse path is encountered the first time, the fuse path is re-created
taking the mount prefix into account. When the file is created for the path
the second time from the cached mount info, the mount prefix is silently
ignored, causing the new file not to be equal to the old one (it is even
invalid):

This change just looks up the mount info in the cache and then continues
constructing the path as it would without having the info cached.

https://bugzilla.gnome.org/show_bug.cgi?id=766294