mtp: Use libusb interruption mechanism
Now that libusb 1.0.21 is finally out, there is a way to explicitly
interrupt a thread that is waiting for events. This allows us to
use a long timeout on our event waiting, which we can interrupt,
rather than a short, one second, timeout. The only nuance is that
we need to link against libusb directly to be able to use the
interrupt mechanism. This is a little bit ugly, but is necessary
unless libmtp wraps the interruption call.