fuse: Add GVFS_DEBUG_FUSE to enable debug output
It was not possible to enable debug output from the fuse daemon without
rebuilding. This patch replaces custom debug handling in the similar way
as it is done in other daemons (using GLib logging functionality).
Environment variable GVFS_DEBUG_FUSE can be specified in order to enable
debugging and SIGUSR2 signal can be sent also to toggle debug output.
The debug output will be printed at stdout, so the debug output can be
seen on gvfsd output thanks to the commit 85bacbf.