afc: Fix GStreamer playback of native files
GStreamer gets very confused when the offset after a seek isn't what it
expected, and told us that the "Stream contains no data." when trying to
play back using giosrc.

But we could play the video just fine going through fuse.

FUSE hides bugs with relative seeks, such as our absolute offset being
wrong when doing absolute seeks. To make sure that the offset we return
is correct, call tell() and use that as the new offset.

https://bugzilla.gnome.org/show_bug.cgi?id=767405