smb: Set always G_FILE_ATTRIBUTE_ACCESS_CAN_WRITE for files
G_FILE_ATTRIBUTE_ACCESS_CAN_WRITE is set for regular files only when
S_IWUSR is not set currently (i.e. FALSE only). It seems to me that
it is reliable to set this to TRUE for regular files in other cases.
Applications should not rely on those attributes, however it is crucial
for some apps which rely on them still (e.g. File Roller).

https://bugzilla.gnome.org/show_bug.cgi?id=724497