smb: Remove maximum read size during reads
The smb read call can take any arbitrary size and the both the
libsmbclient SMB1 and SMB2 engines will break this down into as
many simultaneous on the wire calls as needed to pipeline
the reads / writes.

Restricting this to 65534 will be slow, so remove this
restriction.

See https://lists.samba.org/archive/samba/2016-October/204225.html

https://bugzilla.gnome.org/show_bug.cgi?id=773632