udisks2: Always use generic notification when unmounting
"You can now unplug %s\n" is used for mounts with filesystem,
"%s has been unmounted\n" is used otherwise when unmounting.
It is confusing to say "unplug" for drives with multiple mounts,
or for drives, which need to be ejected, or stopped. I think
"unplug" should be used only when ejecting, or stopping. It might
lead to data loss if user unplug the device after unmounting one
of several mounts.

https://bugzilla.gnome.org/show_bug.cgi?id=774192