recent: Fix crashes when DISPLAY is not set
Recent backend is based on GtkRecentManager and calls gtk_init.
Unfortunatelly, gtk_init aborts with "cannot open display: " warning
if DISPLAY variable is not set. It happens e.g. when recent backend
is spawned from a terminal without a graphical session. Systemd user
units make this even worse and this happens also in graphical session
if gvfs was running before the graphical session starts...

Let's remove gtk_init call. It seems that GtkRecentManager works
properly without it, because the display isn't really needed for it.

https://bugzilla.gnome.org/show_bug.cgi?id=773300