network: Do not spam journal by useless warnings
The following warning is always printed from network backend
if smb backend is installed and network backend is mounted:
Couldn't create directory monitor on smb://[WORKGROUP]/.

You can see it with two different error messages. The first is
printed if we are not in samba environment, or the workgroup is
misconfigured:
Error: The specified location is not mounted

The second is printed in other cases, because monitoring is not
supported by smb backend:
Error: Operation not supported by backend

This isn't really useful and just spams the journal, let's use
g_debug instead of g_warning.